Add support for producing GitHub Releases
This changeset uses a Docker image with Go cross-compilation support.
The `pkg` directory is mounted as a volume within the container so that
binaries can be copied back to the Docker container host.

In addition, the `github-release` command line tool is used within the
`github-release` Make target to automatically produce GitHub Releases
using the `BUILDKITE_TAG` environment variable.