Add support for libpq environment variables
It turns out that the `pq` package already has support for `libpq`
environment variables, but they only take effect if a database
connection string isn't explicitly provided.

This changeset makes the default database URL an empty string, which
allows the environment variables to be used, if set.

Attempts to resolve #3.